Move get_working_csv duplication into catalog util
